% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{convert_format}
\alias{convert_format}
\title{Convert protein column names between formats}
\usage{
convert_format(
  data,
  target_format = c("gene", "seqid_sl", "seqid_dot", "uniprot"),
  id_col = "SampleID",
  age_col = "Age"
)
}
\arguments{
\item{data}{Data frame with protein columns}

\item{target_format}{Target naming convention}

\item{id_col}{Sample ID column name to preserve.}

\item{age_col}{Chronological age column name to preserve.}
}
\value{
Data frame with renamed protein columns
}
\description{
Uses prolist.csv to map between gene symbols, SeqIds, and UniProt.
}
