% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{detect_format}
\alias{detect_format}
\title{Detect the naming convention of protein columns}
\usage{
detect_format(colnames_vec)
}
\arguments{
\item{colnames_vec}{Character vector of column names}
}
\value{
One of "gene", "seqid_sl", "seqid_dot", "uniprot"
}
\description{
Detect the naming convention of protein columns
}
