% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_visualization.R
\name{.clock_outputs_long}
\alias{.clock_outputs_long}
\title{Convert clock outputs to long format}
\usage{
.clock_outputs_long(clock_outputs)
}
\description{
Convert clock outputs to long format
}
\keyword{internal}
