% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goeminne2025.R
\name{goeminne2025_organaging_proteins}
\alias{goeminne2025_organaging_proteins}
\title{List Goeminne 2025 organ aging model proteins}
\usage{
goeminne2025_organaging_proteins()
}
\value{
data.frame with organ, model type, fold, protein, and weight.
}
\description{
List Goeminne 2025 organ aging model proteins
}
