% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuo2024_pac.R
\name{kuo2024_pac}
\alias{kuo2024_pac}
\title{Kuo et al. (2024) PAC Olink proteomic aging clock}
\description{
Kuo CL, Chen Z, Liu P, et al. Proteomic aging clock (PAC) predicts
age-related outcomes in middle-aged and older adults. Aging Cell. 2024.
}
\details{
PAC uses chronological age plus 128 Olink proteins to estimate a
mortality-calibrated proteomic age.
}
