% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuo2024_pac.R
\name{kuo2024_pac_proteins}
\alias{kuo2024_pac_proteins}
\title{List Kuo 2024 PAC predictors}
\usage{
kuo2024_pac_proteins()
}
\value{
data.frame with predictor names and coefficients.
}
\description{
List Kuo 2024 PAC predictors
}
