% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh2023.R
\name{oh2023}
\alias{oh2023}
\title{Oh et al. (2023) conventional proteomic aging clock}
\description{
Oh HS, Rutledge J, et al. Organ aging signatures in the plasma proteome
track health and disease. Nature. 2023;624:164-172.
doi: 10.1038/s41586-023-06802-1
}
\details{
Bagged LASSO ensemble (500 models) on SOMAscan v4 (4,778 proteins),
trained on 1,398 healthy participants from Knight-ADRC.
}
