% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh2023.R
\name{oh2023_conventional_proteins}
\alias{oh2023_conventional_proteins}
\title{List Oh 2023 conventional clock proteins}
\usage{
oh2023_conventional_proteins()
}
\description{
List Oh 2023 conventional clock proteins
}
