% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_utils.R
\name{olink_protein_map}
\alias{olink_protein_map}
\title{List built-in Olink gene-UniProt mappings}
\usage{
olink_protein_map()
}
\value{
data.frame with Olink assay/gene names, UniProt accessions, assay IDs,
  panels, and source.
}
\description{
List built-in Olink gene-UniProt mappings
}
