% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanaka2018.R
\name{tanaka2018}
\alias{tanaka2018}
\title{Tanaka et al. (2018) 76-protein elastic net proteomic clock}
\description{
Tanaka T, et al. Aging Cell 2018;17(5):e12799.
Coefficients from eLife (2020) Supplementary File 1D.
}
