% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuo2024_pac.R
\name{compute_kuo2024_pac_age}
\alias{compute_kuo2024_pac_age}
\title{Compute PAC proteomic age via Kuo et al. (2024)}
\usage{
compute_kuo2024_pac_age(data, id_col = "SampleID", age_col = "Age")
}
\arguments{
\item{data}{data.frame with Olink NPX protein columns plus age.}

\item{id_col}{Sample ID column name.}

\item{age_col}{Chronological age column name.}
}
\value{
data.frame with PAC proteomic age and age acceleration.
}
\description{
Compute PAC proteomic age via Kuo et al. (2024)
}
