% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lehallier2019.R
\name{compute_lehallier2019_age}
\alias{compute_lehallier2019_age}
\title{Compute proteomic age via Lehallier 2019 LASSO clock}
\usage{
compute_lehallier2019_age(
  data,
  id_col = "SampleID",
  age_col = "Age",
  sex_col = "Sex",
  male_value = 0,
  match_by = c("uniprot", "gene", "seqid_sl", "seqid_dot")
)
}
\arguments{
\item{data}{data.frame with protein columns + Age column.}

\item{id_col}{Sample ID column name.}

\item{age_col}{Chronological age column name.}

\item{sex_col}{Sex column name (0=male, 1=female).}

\item{male_value}{Value coding for male.}

\item{match_by}{How to match: \code{"uniprot"} (default), \code{"gene"},
\code{"seqid_sl"}, \code{"seqid_dot"}.}
}
\value{
data.frame with proteomic_age, age_acceleration
}
\description{
Compute proteomic age via Lehallier 2019 LASSO clock
}
