% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oh2023.R
\name{compute_oh2023_conventional_age}
\alias{compute_oh2023_conventional_age}
\title{Compute proteomic age via Oh 2023 conventional clock}
\usage{
compute_oh2023_conventional_age(
  data,
  id_col = "SampleID",
  age_col = "Age",
  sex_col = "Sex_F",
  match_by = c("seqid_dot", "uniprot", "gene", "seqid_sl")
)
}
\arguments{
\item{data}{data.frame with protein columns + Age column.}

\item{id_col}{Sample ID column name.}

\item{age_col}{Chronological age column name.}

\item{sex_col}{Optional sex column name used if present.}

\item{match_by}{How to match: "seqid_dot" (default), "uniprot", "gene", "seqid_sl".}
}
\value{
data.frame with proteomic_age, age_acceleration
}
\description{
Compute proteomic age via Oh 2023 conventional clock
}
