% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang2024.R
\name{compute_wang2024_aric_age}
\alias{compute_wang2024_aric_age}
\title{Compute proteomic age via Wang 2024 ARIC midlife clock}
\usage{
compute_wang2024_aric_age(
  data,
  id_col = "SampleID",
  age_col = "Age",
  match_by = c("seqid_dot", "uniprot", "gene", "seqid_sl")
)
}
\arguments{
\item{data}{data.frame with protein columns + Age.}

\item{id_col}{Sample ID column name.}

\item{age_col}{Chronological age column name.}

\item{match_by}{How to match: "seqid_dot" (default), "uniprot", "gene", "seqid_sl".}
}
\value{
data.frame with proteomic_age, age_acceleration
}
\description{
Compute proteomic age via Wang 2024 ARIC midlife clock
}
